+++ Haze/Haze/Compiler/Source/HazeCompilerArrayValue.cpp
@@ -45,16 +45,16 @@ HazeCompilerArrayValue::~HazeCompilerArrayValue()
 {
 }
 
-uint32 HazeCompilerArrayValue::GetSizeByLevel(uint32 level)
+uint64 HazeCompilerArrayValue::GetSizeByLevel(uint64 level)
 {
-	uint32 ret = 0;
+	uint64 ret = 0;
 	if (level + 1 < m_SizeValues.size())
 	{
 		ret = 1;
 		for (size_t i = level + 1; i < m_SizeValues.size(); i++)
 		{
 			ret *= m_SizeValues[i]->GetValueType().PrimaryType == HazeValueType::UnsignedLong || m_SizeValues[i]->GetValueType().PrimaryType == HazeValueType::Long ?
-				(uint32)m_SizeValues[i]->GetValue().Value.UnsignedLong : m_SizeValues[i]->GetValue().Value.UnsignedInt;
+				m_SizeValues[i]->GetValue().Value.UnsignedLong : m_SizeValues[i]->GetValue().Value.UnsignedInt;
 		}
 	}
